Refactored the function "getNextWednesdayNoonET" so that the date on which the trivia game is played can be easily modified.
Also changed the function's name from "getNextWednesdayNoonET" to "getNextTriviaGameDate" to better summarize
that the function is giving the next date of the trivia game instead of just giving the next Wednesday Noon ET Date.

Modified the "getGameStatus" function to remove a hard coded number and put it in a variable.